ERDs V1 and V2, and the Hierarchy Diagram
V1 of my ERD had a lot of objects that should have been combined together with different record types, so I went through and updated that, as well as changed some master-detail relationships and junction objects around. The final version is mostly similar to V2, with the exception of the removal of the R. Assignment junction object and the master-detail between Account and Request being converted to a lookup detail.